requirements.txt
Django==6.0
djangorestframework==3.17.1
django-cors-headers==4.9.0
dj-database-url==3.1.2
gunicorn==25.3.0
python-dotenv==1.2.2
